+++ packages/diagram-ui/src/features/canvas-content-edit/handler/noopMoveHandler.ts
@@ -1,0 +1,16 @@
+import { Edit, TransactionalAction } from "@hylimo/diagram-protocol";
+import { MoveHandler } from "../../move/moveHandler.js";
+
+/**
+ * Move handler that does nothing
+ * Can be used as fallback if no move handler is available
+ */
+export class NoopMoveHandler extends MoveHandler {
+    override generateActions(): TransactionalAction[] {
+        return [];
+    }
+
+    override generateEdits(): Edit[] {
+        return [];
+    }
+}
