+++ packages/diagram-ui/src/features/canvas-content-edit/handler/resizeHandler.ts
@@ -1,6 +1,7 @@
-import { DefaultEditTypes, Math2D } from "@hylimo/diagram-common";
-import { MoveHandler } from "./moveHandler.js";
+import { DefaultEditTypes } from "@hylimo/diagram-common";
 import { Edit, ResizeEdit } from "@hylimo/diagram-protocol";
+import { Matrix } from "transformation-matrix";
+import { MoveHandler } from "../../move/moveHandler.js";
 
 /**
  * Elements with an optional original width and height.
@@ -16,47 +17,39 @@ export interface ElementsGroupedBySize {
 
 /**
  * A move handler that resizes the elements.
+ * Expects relative coordinates to its own coordinate system.
  */
 export class ResizeHandler extends MoveHandler {
     /**
-     * The rotation of the primary resize element, in radians.
-     */
-    private readonly rotation: number;
-
-    /**
      * Creates a new ResizeHandler.
      *
-     * @param transactionId The transaction id.
-     * @param rotation the rotation of the primary resize element, in degrees.
      * @param scaleX the scale factor in x direction. The resize is scaled by this to account for resize in the opposite direction.
      * @param scaleY the scale factor in y direction. The resize is scaled by this to account for resize in the opposite direction.
      * @param originalWidth the original width of the primary resize element, used to calculate the resize factor.
      * @param originalHeight the original height of the primary resize element, used to calculate the resize factor.
      * @param groupedElements the elements grouped by size.
+     * @param transformationMatrix the transformation matrix to apply to obtain the relative position.
      */
     constructor(
-        transactionId: string,
-        rotation: number,
         private readonly scaleX: number | undefined,
         private readonly scaleY: number | undefined,
         private readonly originalWidth: number,
         private readonly originalHeight: number,
-        private readonly groupedElements: ElementsGroupedBySize[]
+        private readonly groupedElements: ElementsGroupedBySize[],
+        transformationMatrix: Matrix
     ) {
-        super(transactionId);
-        this.rotation = rotation * (Math.PI / 180);
+        super(transformationMatrix);
     }
 
-    protected override generateEdits(dx: number, dy: number): Edit[] {
+    override generateEdits(x: number, y: number): Edit[] {
         let factorX: number | undefined = undefined;
         let factorY: number | undefined = undefined;
-        const normalizedDelta = Math2D.rotate({ x: dx, y: dy }, -this.rotation);
 
         if (this.scaleX !== undefined) {
-            factorX = Math.abs((normalizedDelta.x * this.scaleX + this.originalWidth) / this.originalWidth);
+            factorX = Math.abs((x * this.scaleX + this.originalWidth) / this.originalWidth);
         }
         if (this.scaleY !== undefined) {
-            factorY = Math.abs((normalizedDelta.y * this.scaleY + this.originalHeight) / this.originalHeight);
+            factorY = Math.abs((y * this.scaleY + this.originalHeight) / this.originalHeight);
         }
         const edits: Edit[] = [];
         for (const group of this.groupedElements) {
