+++ packages/diagram-ui/src/features/canvas-content-edit/handler/translationMoveHandler.ts
@@ -1,5 +1,5 @@
 import { Edit, MoveEdit } from "@hylimo/diagram-protocol";
-import { MoveHandler } from "./moveHandler.js";
+import { MoveHandler } from "../../move/moveHandler.js";
 import { DefaultEditTypes } from "@hylimo/diagram-common";
 import { Matrix, applyToPoint } from "transformation-matrix";
 
@@ -18,27 +18,28 @@ export interface ElementsGroupedByTransformation {
 }
 
 /**
- * Move handler for translations of absolute and relative points
+ * Move handler for translations of absolute and relative points.
+ * Expects relative coordinates in the root canvas coordinate system.
  */
 export class TranslationMoveHandler extends MoveHandler {
     /**
      * Creats a new TranslateMovehandler
      *
      * @param elements the ids of the points to move
-     * @param transactionId the id of the transaction
+     * @param transformationMatrix the transformation matrix to apply to obtain the relative position
      */
     constructor(
         readonly elements: ElementsGroupedByTransformation[],
-        transactionId: string
+        transformationMatrix: Matrix
     ) {
-        super(transactionId);
+        super(transformationMatrix);
     }
 
-    protected override generateEdits(dx: number, dy: number, event: MouseEvent): Edit[] {
-        let offsetX = dx;
-        let offsetY = dy;
+    override generateEdits(x: number, y: number, event: MouseEvent): Edit[] {
+        let offsetX = x;
+        let offsetY = y;
         if (event.shiftKey) {
-            if (Math.abs(dx) > Math.abs(dy)) {
+            if (Math.abs(x) > Math.abs(y)) {
                 offsetY = 0;
             } else {
                 offsetX = 0;
