+++ packages/diagram-ui/src/features/canvas-content-edit/movedElementsSelector.ts
