+++ packages/diagram-ui/src/features/canvas-content-edit/splitCanvasSegmentMouseListener.ts
