--- packages/diagram-ui/src/features/split-canvas-segment/di.config.ts
@@ -1,11 +1,0 @@
-import { ContainerModule } from "inversify";
-import { TYPES } from "sprotty";
-import { SplitCanvasSegmentMouseListener } from "./splitCanvasSegmentMouseListener.js";
-
-/**
- * Module for splitting canvas connection segments into two parts using an edit
- */
-export const splitCanvasSegmentModule = new ContainerModule((bind, _unbind, _isBound) => {
-    bind(SplitCanvasSegmentMouseListener).toSelf().inSingletonScope();
-    bind(TYPES.MouseListener).toService(SplitCanvasSegmentMouseListener);
-});
