+++ packages/diagram-ui/src/index.ts
@@ -8,30 +8,29 @@ export * from "./features/animation/model.js";
 export * from "./features/canvas-bounds/di.config.js";
 export * from "./features/canvas-bounds/resetCanvasBoundsAction.js";
 export * from "./features/canvas-bounds/resetCanvasBoundsCommand.js";
+export * from "./features/canvas-content-edit/handler/axisAlignedSegmentEditHandler.js";
+export * from "./features/canvas-content-edit/handler/lineMoveHandler.js";
+export * from "./features/canvas-content-edit/handler/noopMoveHandler.js";
+export * from "./features/canvas-content-edit/handler/resizeHandler.js";
+export * from "./features/canvas-content-edit/handler/rotationHandler.js";
+export * from "./features/canvas-content-edit/handler/translationMoveHandler.js";
+export * from "./features/canvas-content-edit/di.config.js";
+export * from "./features/canvas-content-edit/moveEditCanvasContentMouseListener.js";
+export * from "./features/canvas-content-edit/movedElementsSelector.js";
+export * from "./features/canvas-content-edit/splitCanvasSegmentMouseListener.js";
 export * from "./features/config/configManager.js";
 export * from "./features/config/di.config.js";
 export * from "./features/config/editorConfigUpdatedCommand.js";
-export * from "./features/create-move/createAndMoveAction.js";
-export * from "./features/create-move/createAndMoveCommand.js";
-export * from "./features/create-move/createAndMoveMouseListener.js";
-export * from "./features/create-move/createElementMoveHandler.js";
-export * from "./features/create-move/createMoveHandler.js";
-export * from "./features/create-move/di.config.js";
 export * from "./features/layout/positionProvider.js";
-export * from "./features/move/axisAlignedSegmentEditHandler.js";
 export * from "./features/move/di.config.js";
-export * from "./features/move/lineMoveHandler.js";
 export * from "./features/move/moveHandler.js";
 export * from "./features/move/moveMouseListener.js";
-export * from "./features/move/movedElementsSelector.js";
-export * from "./features/move/resizeHandler.js";
-export * from "./features/move/rotationHandler.js";
-export * from "./features/move/translationMoveHandler.js";
+export * from "./features/move/transactionalMoveAction.js";
+export * from "./features/move/transactionalMoveCommand.js";
 export * from "./features/navigation/di.config.js";
 export * from "./features/navigation/navigationMouseListener.js";
-export * from "./features/split-canvas-segment/di.config.js";
-export * from "./features/split-canvas-segment/splitCanvasSegmentMouseListener.js";
 export * from "./features/toolbox/ci.config.js";
+export * from "./features/toolbox/createElementMoveHandler.js";
 export * from "./features/toolbox/toolbox.js";
 export * from "./features/transaction/di.config.js";
 export * from "./features/transaction/transactionIdProvider.js";
